requirements.txt
streamlit
requests
pandas
numpy
scikit-learn
scipy
plotly
google-genai
